requirements.txt
# Core data processing
pandas>=2.0.0

# Templating
jinja2>=3.1.0

# Web scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# PDF export
Pillow>=10.0.0
